+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/BlobQuerySerialization.java
@@ -5,7 +5,11 @@ package com.azure.storage.blob.models;
 
 /**
  * Defines the input and output serialization for a blob quick query request.
- * either {@link BlobQueryJsonSerialization} or {@link BlobQueryDelimitedSerialization}
+ * either {@link BlobQueryJsonSerialization}, {@link BlobQueryDelimitedSerialization}, or
+ * {@link BlobQueryArrowSerialization}.
+ * <p>
+ * Note: {@link BlobQueryArrowSerialization} can only be used as an output.
+ * </p>
  */
 public interface BlobQuerySerialization {
 }