+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.3.0-beta.1 (Unreleased)
 - Added support for the 2019-02-10 service version.
 - Added support to schedule file expiration. 
+- Added support to specify Arrow Output Serialization when querying a file. 
 
 ## 12.2.0 (2020-08-13)
 - Fixed bug where Query Input Stream would throw when a ByteBuffer of length 0 was encountered.